Cloning the base code of the project on Github, installing the dependencies, and setting up the application breakpoints
Obtaining the project template from Github and performing the cloning of the project code. Installing the project dependencies defined in the 'package.json' file. Finally, preparing the environment to make the application responsive by creating the 'deviceBreakpoints.js' file, which stores an object containing the breakpoints that will be used in the application to make the project accessible on different screen sizes