Making the 'Priorities', 'ShortTask', 'Filter', and 'Task' components responsive and creating buttons to close and display the sidebar
Continuing the process of making the application responsive by rendering the 'Priorities', 'ShortTask', 'Filter', and 'Task' components responsive. Also, adjusting the main page title present in the 'Header' component using the 'rem' unit instead of pixels to define the font size. Finally, restoring the display of the side menu on smaller screens, making it occupy the entire available screen. Additionally, creating a button in the 'Header' component to show the menu and a button in the 'SideMenu' component, which represents the menu itself, to close it