Creating the logic for displaying the menu, applying an animation to the menu, preparing the application for screens smaller than 280 pixels, and finalizing the development
Implementing the logic for displaying the menu, applying styles to the menu through 'data attributes,' storing information about whether the menu is open or not in the 'menuIsOpen' state created within the 'Home' component, passing this information to the 'SideMenu' component, which represents the side menu. Using the buttons present in the 'Header' component and the 'SideMenu' component to open and close the side menu, respectively, changing the value of the 'data attribute' responsible for displaying the menu or not through the function to update the 'menuIsOpen' state. Updating the 'deviceBreakpoints.js' file containing the application's breakpoints, using the value of '280px' for the smallest breakpoint and '320px' for the second smallest. Finally, completing the development by adapting the application for screens equal to or smaller than 280 pixels wide, changing the font size of the main page title in the 'Header' component to '1.5rem' and reducing the size of the button and icons present in the 'Header'